Remove unnecessary static_cast

+++ src/fourier_transform.h
@@ -65,9 +65,7 @@ inline Expression fourier_transform_operator(Operator op, const DynamicIndex& in
     coefficient = std::exp(coefficient) * normalization;
 
     Operator transformed_op(op.type(), op.spin(), k);
-    result += Term(Term::complex_type(static_cast<float>(coefficient.real()),
-                                      static_cast<float>(coefficient.imag())),
-                   {transformed_op});
+    result += Term(Term::complex_type(coefficient.real(), coefficient.imag()), {transformed_op});
   }
 
   return result;
